Update JobService to use security context

+++ backend/src/main/java/com/jobtracker/service/JobService.java
@@ -2,11 +2,13 @@ package com.jobtracker.service;
 
 import com.jobtracker.entity.User;
 import com.jobtracker.exception.JobNotFoundException;
+import com.jobtracker.exception.UserForbiddenJobException;
 import com.jobtracker.repository.JobRepository;
 import com.jobtracker.dto.JobRequestDTO;
 import com.jobtracker.dto.JobResponseDTO;
 import com.jobtracker.entity.Job;
 import com.jobtracker.repository.UserRepository;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -16,11 +18,9 @@ import java.util.Optional;
 public class JobService {
 
     private final JobRepository jobRepository;
-    private final UserRepository userRepository;
 
-    public JobService(JobRepository jobRepository, UserRepository userRepository) {
+    public JobService(JobRepository jobRepository) {
         this.jobRepository = jobRepository;
-        this.userRepository = userRepository;
     }
 
     private JobResponseDTO toJobResponseDTO(Job job) {
@@ -36,9 +36,6 @@ public class JobService {
 
     private Job toJob (JobRequestDTO jobRequest) {
         Job job = new Job();
-        User user = userRepository.findById(jobRequest.getUserId())
-                .orElseThrow(() -> new RuntimeException("User not found with id: " + jobRequest.getUserId()));
-        job.setUser(user);
         job.setCompanyName(jobRequest.getCompanyName());
         job.setJobTitle(jobRequest.getJobTitle());
         job.setStatus(jobRequest.getStatus());
@@ -47,14 +44,17 @@ public class JobService {
     }
 
     public List<JobResponseDTO> getAllJobs() {
-        return jobRepository.findAll()
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        return jobRepository.findByUser(user)
                             .stream()
                             .map(this::toJobResponseDTO)
                             .toList();
     }
 
     public JobResponseDTO postJob(JobRequestDTO jobRequest) {
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
         Job job = toJob(jobRequest);
+        job.setUser(user);
         Job savedJob = jobRepository.save(job);
         return toJobResponseDTO(jobRepository.findById(savedJob.getId()).orElseThrow());
     }
@@ -65,12 +65,24 @@ public class JobService {
         if (tempJob.isEmpty()) {
             throw new JobNotFoundException(id);
         }
+
+        User currentUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!tempJob.get().getUser().getId().equals(currentUser.getId())) {
+            throw new UserForbiddenJobException();
+        }
+
         return toJobResponseDTO(tempJob.get());
     }
 
     public void deleteJob(Long id) {
         Job job = jobRepository.findById(id)
                 .orElseThrow(() -> new JobNotFoundException(id));
+
+        User currentUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!job.getUser().getId().equals(currentUser.getId())) {
+            throw new UserForbiddenJobException();
+        }
+
         jobRepository.delete(job);
     }
 }
